Expiracion de mensajes.

+++ src/main/java/redis/rmq/Nest.java
@@ -64,9 +64,9 @@ public class Nest {
         return string;
     }
 
-    public Integer incr() {
+    public Long incr() {
         Jedis jedis = getResource();
-        Integer incr = jedis.incr(key());
+        Long incr = jedis.incr(key());
         returnResource(jedis);
         return incr;
     }
@@ -78,16 +78,16 @@ public class Nest {
         return multi;
     }
 
-    public Integer del() {
+    public Long del() {
         Jedis jedis = getResource();
-        Integer del = jedis.del(key());
+        Long del = jedis.del(key());
         returnResource(jedis);
         return del;
     }
 
-    public Integer exists() {
+    public Long exists() {
         Jedis jedis = getResource();
-        Integer exists = jedis.exists(key());
+        Long exists = jedis.exists(key());
         returnResource(jedis);
         return exists;
     }
@@ -114,16 +114,16 @@ public class Nest {
         return value;
     }
 
-    public Integer hdel(String field) {
+    public Long hdel(String field) {
         Jedis jedis = getResource();
-        Integer hdel = jedis.hdel(key(), field);
+        Long hdel = jedis.hdel(key(), field);
         returnResource(jedis);
         return hdel;
     }
 
-    public Integer hlen() {
+    public Long hlen() {
         Jedis jedis = getResource();
-        Integer hlen = jedis.hlen(key());
+        Long hlen = jedis.hlen(key());
         returnResource(jedis);
         return hlen;
     }
@@ -136,16 +136,16 @@ public class Nest {
     }
 
     // Redis Set Operations
-    public Integer sadd(String member) {
+    public Long sadd(String member) {
         Jedis jedis = getResource();
-        Integer reply = jedis.sadd(key(), member);
+        Long reply = jedis.sadd(key(), member);
         returnResource(jedis);
         return reply;
     }
 
-    public Integer srem(String member) {
+    public Long srem(String member) {
         Jedis jedis = getResource();
-        Integer reply = jedis.srem(key(), member);
+        Long reply = jedis.srem(key(), member);
         returnResource(jedis);
         return reply;
     }
@@ -158,9 +158,9 @@ public class Nest {
     }
 
     // Redis List Operations
-    public Integer rpush(String string) {
+    public Long rpush(String string) {
         Jedis jedis = getResource();
-        Integer rpush = jedis.rpush(key(), string);
+        Long rpush = jedis.rpush(key(), string);
         returnResource(jedis);
         return rpush;
     }
@@ -179,16 +179,16 @@ public class Nest {
         return lindex;
     }
 
-    public Integer llen() {
+    public Long llen() {
         Jedis jedis = getResource();
-        Integer llen = jedis.llen(key());
+        Long llen = jedis.llen(key());
         returnResource(jedis);
         return llen;
     }
 
-    public Integer lrem(int count, String value) {
+    public Long lrem(int count, String value) {
         Jedis jedis = getResource();
-        Integer lrem = jedis.lrem(key(), count, value);
+        Long lrem = jedis.lrem(key(), count, value);
         returnResource(jedis);
         return lrem;
     }
@@ -224,16 +224,16 @@ public class Nest {
         return zrange;
     }
 
-    public Integer zadd(float score, String member) {
+    public Long zadd(float score, String member) {
         Jedis jedis = getResource();
-        Integer zadd = jedis.zadd(key(), score, member);
+        Long zadd = jedis.zadd(key(), score, member);
         returnResource(jedis);
         return zadd;
     }
 
-    public Integer zcard() {
+    public Long zcard() {
         Jedis jedis = getResource();
-        Integer zadd = jedis.zcard(key());
+        Long zadd = jedis.zcard(key());
         returnResource(jedis);
         return zadd;
     }
@@ -266,9 +266,9 @@ public class Nest {
         return result;
     }
 
-    public Integer publish(String message) {
+    public Long publish(String message) {
         Jedis jedis = getResource();
-        Integer result = jedis.publish(key(), message);
+        Long result = jedis.publish(key(), message);
         returnResource(jedis);
         return result;
     }
